separate provider in module to provide better access to separated areas
